integer compress

+++ libsuru-vm/include/suru/vm/codeunit.hpp
@@ -11,11 +11,11 @@ namespace suru::vm {
 
 struct Chunk {
     std::string name;
-    std::size_t code_begin {0};
-    std::size_t code_end {0};
-    std::size_t arity {0};
-    std::size_t slots {0};
-    std::size_t upvalues {0};
+    std::uint32_t code_begin {0};
+    std::uint32_t code_end {0};
+    std::uint8_t arity {0};
+    std::uint8_t slots {0};
+    std::uint8_t upvalues {0};
 };
 
 struct CodeUnit {
